+++ Implementations/FAnsi.Implementations.MySql/MySqlQuerySyntaxHelper.cs
@@ -14,6 +14,12 @@ namespace FAnsi.Implementations.MySql
         public override int MaximumTableLength => 64;
         public override int MaximumColumnLength => 64;
 
+
+        
+        public override string OpenQualifier => "`";
+
+        public override string CloseQualifier => "`";
+
         public MySqlQuerySyntaxHelper() : base(new MySqlTypeTranslater(), new MySqlAggregateHelper(),new MySqlUpdateHelper(),DatabaseType.MySql)//no specific type translation required
         {
         }
@@ -35,15 +41,21 @@ namespace FAnsi.Implementations.MySql
         /// <returns></returns>
         private string GetRuntimeNameWithDoubledBackticks(string s)
         {
-            return GetRuntimeName(s).Replace("`","``");
+            return GetRuntimeName(s)?.Replace("`","``");
         }
+
+        protected override string UnescapeWrappedNameBody(string name)
+        {
+            return name.Replace("``","`");
+        }
+
         public override string EnsureFullyQualified(string databaseName, string schema, string tableName)
         {
             //if there is no schema address it as db..table (which is the same as db.dbo.table in Microsoft SQL Server)
             if (!string.IsNullOrWhiteSpace(schema))
                 throw new NotSupportedException("Schema (e.g. .dbo. not supported by MySql)");
 
-            return "`" + GetRuntimeNameWithDoubledBackticks(databaseName) + "`" + DatabaseTableSeparator + "`" + GetRuntimeNameWithDoubledBackticks(tableName) + "`";
+            return  EnsureWrapped(databaseName) + DatabaseTableSeparator + EnsureWrapped(tableName);
         }
 
         public override TopXResponse HowDoWeAchieveTopX(int x)